Implement __replace__ on 3.13
Fixes #1313
