+++ .readthedocs.yaml
@@ -5,7 +5,7 @@ build:
   os: ubuntu-lts-latest
   tools:
     # Keep version in sync with tox.ini/docs and ci.yml/docs.
-    python: "3.12"
+    python: "3.13"
   jobs:
     # Need the tags to calculate the version (sometimes).
     post_checkout:
