+++ tox.ini
@@ -62,7 +62,7 @@ commands = pytest --codspeed -n auto bench/test_benchmarks.py
 
 [testenv:docs]
 # Keep base_python in-sync with ci.yml/docs and .readthedocs.yaml.
-base_python = py312
+base_python = py313
 extras = docs
 commands =
     sphinx-build -n -T -W -b html -d {envtmpdir}/doctrees docs docs/_build/html
